move handshake to consensus package
